Move TechdocsContainerRunner interface from publish to generate where it is used
