+++ .changeset/big-garlics-punch.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-techdocs-node': patch
+---
+
+Move `TechdocsContainerRunner` from `publish` to `generate`.
