+++ plugins/techdocs-node/src/stages/publish/index.ts
@@ -24,5 +24,4 @@ export type {
   MigrateRequest,
   ReadinessResponse,
   TechDocsMetadata,
-  TechDocsContainerRunner,
 } from './types';
